Refactoring: Merge OperationStatusHelper into DBSQLOperation
